feat: declare content in wallet.json
