+++ src/components/Wallet/BigImageCard/styles.module.css
@@ -1,5 +1,4 @@
 .card {
-  background: var(--mui-palette-border-background);
   outline: 1px solid var(--mui-palette-border-light);
   border-radius: 16px;
   padding: 16px;
@@ -13,6 +12,11 @@
 
 .image {
   width: 240px;
+}
+
+.title {
+  width: 80%;
+  margin-bottom: 24px;
 }
 
 @media (min-width: 600px) {
