+++ src/pages/wallet.tsx
@@ -1,0 +1,6 @@
+import type { NextPage } from 'next'
+import { Wallet } from '@/components/Wallet'
+
+const WalletPage: NextPage = () => <Wallet />
+
+export default WalletPage
